+++ internal/plugin/host/plugin.go
@@ -14,7 +14,7 @@ type Plugin struct {
 
 // NewPlugin creates a new in memory Plugin assigned to the global scope.
 // Name, Description are the only allowed option. All other options are ignored.
-func NewPlugin(pluginName string, idPrefix string, opt ...Option) *Plugin {
+func NewPlugin(pluginName, idPrefix string, opt ...Option) *Plugin {
 	opts := getOpts(opt...)
 	p := &Plugin{
 		Plugin: &store.Plugin{
@@ -33,7 +33,7 @@ func (c *Plugin) TableName() string {
 	if c.tableName != "" {
 		return c.tableName
 	}
-	return "host_plugin"
+	return "plugin_host"
 }
 
 // SetTableName sets the table name. If the caller attempts to